Articles model tumbnail field change.
Preivously an models.FileField
Now models.ImageField

Reason for change:
Previously possible to upload any file.
Problems with Content-Header when uploading new images if
its not explicitly stated in HTTP header that the content
is an image. Gambled that this will make it impossible.
Maybe won't restrict other files, such as HEIC files etc.

If this does not work, change MIME-types in nginx to also interpret
HEIC files as Content-Type: image